test: migrate DataLocationList to vitest
